LayoutTests/js/dom/modules/import-execution-order-expected.txt

Test import execution order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS count++ is 0
PASS count++ is 1
PASS count++ is 2
PASS count++ is 3
PASS count++ is 4
PASS count++ is 5
PASS count++ is 6
PASS successfullyParsed is true

TEST COMPLETE
